Refactor album services and handlers, and add new album-related tests, services, and handlers

+++ server/src/models/albumImages.model.js
@@ -1,12 +1,46 @@
 const prisma = require("@config/db.config");
 
+const createAlbumImageLink = async (data) => {
+  return await prisma.album_images.create({ data });
+};
+
+const fetchAlbumImage = async (where) => {
+  return await prisma.album_images.findFirst({
+    where,
+    include: {
+      images: true,
+    },
+  });
+};
+
+const fetchAlbumImages = async (where) => {
+  return await prisma.album_images.findMany({
+    where,
+    include: {
+      images: true,
+    },
+  });
+};
+
 const deleteLinksByAlbumId = async (albumId) => {
+  if (!albumId) {
+    throw new Error("Album ID is required");
+  }
+  const albumLinks = await fetchAlbumImages({ album_id: albumId });
+  if (!albumLinks) {
+    throw new Error("No album links found for the given album ID");
+  }
+
   return await prisma.album_images.deleteMany({
     where: {
       album_id: albumId,
     },
   });
 };
+
 module.exports = {
+  createAlbumImageLink,
+  fetchAlbumImage,
+  fetchAlbumImages,
   deleteLinksByAlbumId,
 };
